[update] keep Trace objects in a tree structure
Whenever a promise resolves we copy the ancestor `Trace.stackMap` into
the current async context's `Trace`. We also recursively copy it into
all descendants of the current Trace.

This means any Error stacks that are extended from this point on will
contain the new frames.

Note this leaks lots of memory, see the next commit for some limits.